requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0
paramiko>=3.4.0
python-dotenv>=1.0.0
requests>=2.31.0
